fix integer overflow in udelay and clock_delay

+++ mos/arch/msp430/udelay_hal.h
@@ -30,9 +30,9 @@
     (defined(__GNUC__) && __GNUC__ >= 4 && __GNUC_MINOR__ >= 6)
 
 #include "intrinsics.h"
-#define udelay(x) __delay_cycles(x * CPU_MHZ)
+#define udelay(x) __delay_cycles((uint32_t) x * CPU_MHZ)
 #define mdelay(x) __delay_cycles((uint32_t) x * CPU_MHZ * 1000ul)
-#define clock_delay(x) __delay_cycles(x * 4)
+#define clock_delay(x) __delay_cycles((uint32_t) x * 4)
 
 #else // Old GCC
